[models] Added device model (BACKWARD INCOMPATIBLE) #52
This backward incompatible change modifies the database
and model logic in few important ways:

A new Device model stores information that is specific to
network devices: mac address, name, model, operating system,
while configuration information is optional and delegated to
Config, which does not store mac address and hostname anymore.

This change allows to create users which can operate on device
objects without having permission to edit their configuration,
and also allows the possibility to implement drafting configuration
changes before applying them.

It's a risky change, but I believe it's a good long term investment;
it's also better to do this kind of change now when the OpenWISP
project still has a small userbase than lateri on when it will grow to
thousands of users.